refact: regard our pattern lang as DPL==dataflow pattern language and set new structure
